Refactorización con el patrón Factory
