+++ app/blueprints/recognition/__init__.py
@@ -1,0 +1,1 @@
+# Este archivo permite que Python reconozca este directorio como un paquete
