+++ requirements.txt
@@ -1,11 +1,56 @@
 https://github.com/jloh02/dlib/releases/download/v19.22/dlib-19.22.99-cp310-cp310-win_amd64.whl
-opencv-python
-ultralytics
-face-recognition
-numpy
-flask
-Pillow
-torch
-facenet-pytorch
-torchvision
-scikit-learn
+blinker==1.9.0
+certifi==2025.1.31
+charset-normalizer==3.4.1
+click==8.1.8
+cmake==3.31.6
+colorama==0.4.6
+contourpy==1.3.1
+cycler==0.12.1
+face-recognition==1.3.0
+face-recognition-models==0.3.0
+facenet-pytorch==2.6.0
+filelock==3.18.0
+Flask==3.1.0
+fonttools==4.56.0
+fsspec==2025.3.0
+idna==3.10
+imageio==2.37.0
+itsdangerous==2.2.0
+Jinja2==3.1.6
+joblib==1.4.2
+kiwisolver==1.4.8
+lazy_loader==0.4
+MarkupSafe==3.0.2
+matplotlib==3.10.1
+mpmath==1.3.0
+networkx==3.4.2
+numpy==1.26.4
+opencv-python==4.11.0.86
+packaging==24.2
+pandas==2.2.3
+pillow==10.2.0
+psutil==7.0.0
+py-cpuinfo==9.0.0
+pyparsing==3.2.1
+python-dateutil==2.9.0.post0
+pytz==2025.1
+PyYAML==6.0.2
+requests==2.32.3
+scikit-image==0.25.2
+scikit-learn==1.6.1
+scipy==1.15.2
+seaborn==0.13.2
+six==1.17.0
+sympy==1.13.1
+threadpoolctl==3.6.0
+tifffile==2025.3.13
+torch==2.2.2
+torchvision==0.17.2
+tqdm==4.67.1
+typing_extensions==4.12.2
+tzdata==2025.1
+ultralytics==8.3.94
+ultralytics-thop==2.0.14
+urllib3==2.3.0
+Werkzeug==3.1.3
